move xas_viewer gui to main larch lib
