+++ larch/wxxas/decompose_panel.py
@@ -23,7 +23,7 @@ from larch.wxlib import (BitmapButton, FloatCtrl, get_icon, SimpleText,
                          pack, Button, HLine, Choice, Check, CEN, RCEN,
                          LCEN, Font)
 
-from larch_plugins.xasgui.taskpanel import TaskPanel
+from .taskpanel import TaskPanel
 from larch.math.lincombo_fitting import get_arrays
 
 np.seterr(all='ignore')
